Introduced 'TM_INTERACTIVE_INPUT' env var which controls when the requester is used. The usage of this lib has changed now in that it will be loaded for everything running under TM (not on opt-in).
The new env var can specify 3 modes; ALWAYS, AUTO and NEVER.

• ALWAYS - ignore any existing data on stdin and always request from user
• AUTO - only request input from user when stdin is empty
• NEVER - effectively disengage

(If the value is neither of those, AUTO is assumed)

This change removes the need for the 'helpers' so they are gone (tests updated).

The docs have also been removed, users of this will now have very little to do to use this so probably some words in the TM manual will be more effective.



git-svn-id: http://svn.textmate.org/trunk/Review/Tools/tmread@9256 dfb7d73b-c2ec-0310-8fea-fb051d288c6d