• Implemented tracking of which file descriptors are STDIN.
To achieve this we overload dup() and close(). In those implementations we pass the file descriptors in question to the methods stdin_fd_tracker_did_dup() and stdin_fd_tracker_did_close(). The methods in stdin_fd_tracker.c use a global integer set (intset.c) to maintain the list of STDIN file descriptors.

When a read is requested, we now see if they file descriptor the read is for is really STDIN by calling stdin_fd_tracker_is_stdin().

This change makes the test committed as revision 9344 pass. 

git-svn-id: http://svn.textmate.org/trunk/Tools/tm_interactive_input@9345 dfb7d73b-c2ec-0310-8fea-fb051d288c6d