Rewrite remove_from_intset.
The old remove_from_intset would decrease size and create a copy of the integer set with the element removed, but would not update the type to point to this new copy, so in practice remove_from_intset() would always remove the last element, regardless of argument passed in.

New version uses memmove() to handle removal in-place.


git-svn-id: http://svn.textmate.org/trunk/Tools/tm_interactive_input@10532 dfb7d73b-c2ec-0310-8fea-fb051d288c6d