• Use the system implementation of the actual variant of the system function that was requested.
e.g. If the client app is actually linking against ‘_write$NOCANCEL$UNIX2003’, we now call that implementation after our custom stuff. Previously we just called ‘_write’ regardless.

Also provided system_* functions for functions that we overwrite so we contain the logic of getting to the system implementation in one place.

git-svn-id: http://svn.textmate.org/trunk/Tools/tm_interactive_input@10533 dfb7d73b-c2ec-0310-8fea-fb051d288c6d